requirements.txt
streamlit
groq
pdfplumber
python-dotenv